Make IMDS unavailable in controller config tests (#336)
The code under test falls back to reading AWS region, account ID,
and VPC ID from Instance Metadata Service. The tests currently
assume that this service is unavaiable. This means they fail
when run in environments where this service is available to tests.

This commit fixes this issue by making the tests use a mock
Instance Metadata Service which returns an error. This ensures
that these tests work as expected regardless of whether Instance
Metadata Service is available/reachable.